P4/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conflict-aware-rag-p4"
version = "0.2.0"
description = "Backend API for Conflict-Aware RAG (P4): orchestrates P3 retrieval, P1 NLI, P2 conflict typing, and P6 answer planning."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "pydantic>=2.7.0,<3.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "openai>=1.58.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "httpx>=0.27.0,<0.28.0",
    # Local submodules (must be installed as editable from parent project)
    # "p1 @ file://../P1",
    # "p2 @ file://../P2",
    # "p3 @ file://../P3",
    # "p6 @ file://../P6",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "requests>=2.31.0",
]

[project.scripts]
p4ctl = "p4ctl:main"

[tool.setuptools]
packages = []
# Note: src/ is used but setuptools picks it automatically via find:
# [tool.setuptools.packages.find]
# where = ["src"]